iOS-#67 [계산기] Layout 관련 함수 분리 (CalculatorViewController + layout.swift)
